Notify when active pane changes in control mode
Now, whenever the active pane in a window changes, if the client is a
control mode, a notification will be sent in the following format:

%active-pane-changed @windowid %paneid

This allows control mode clients to display which pane tmux considers to
be active without having to attempt to replicate tmux's rules around
which pane is active.

This notification is also sent following a pane being killed with C-d.
It also sends when a new window joins.